rlp: fix encoding 0 byte. refactor type helpers
The 0 byte is 0x80 in every other rlp encoding lib. Even though the
spec doesn't call this out specifically, it should be similar to int 0
and the empty string.
